+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+### Changed
+
+* Head tracker is now working on any Substreams supported network.
+
+* Removed the need to have a Firehose instance to track head.
+
 ## v1.0.0
 
 * Initial commit